Split grpc-java only benchmark
